support full transform (#300)
* support full transform

* remove commented out code

* small clean up

* more comments

+++ constraintlayout/core/src/main/java/androidx/constraintlayout/core/parser/CLElement.java
@@ -152,7 +152,7 @@ public class CLElement {
 
     public float getFloat() {
         if (this instanceof CLNumber) {
-            return ((CLNumber) this).getInt();
+            return ((CLNumber) this).getFloat();
         }
         return Float.NaN;
     }
